+++ src/mplchart/primitives/barplot.py
@@ -50,7 +50,7 @@ class BarPlot(BindingPrimitive):
     def apply_to_chart(self, chart):
         ax = chart.get_axes()
 
-        result = chart.calc_result(self.indicator)
+        result = chart.calc_result(self.required_indicator())
 
         series = series_data(result, self.item)
 
@@ -62,5 +62,5 @@ class BarPlot(BindingPrimitive):
             alpha=self.alpha,
         )
 
-        xv, yv = chart.mapper.series_xy(series)
+        xv, yv = chart.series_xy(series)
         ax.bar(xv, yv, label=label, **kwargs)