+++ src/mplchart/primitives/lineplot.py
@@ -66,7 +66,7 @@ class LinePlot(BindingPrimitive):
     def apply_to_chart(self, chart):
         ax = chart.get_axes()
 
-        result = chart.calc_result(self.indicator)
+        result = chart.calc_result(self.required_indicator())
 
         series = series_data(result, self.item)
 
@@ -80,7 +80,7 @@ class LinePlot(BindingPrimitive):
             alpha=self.alpha,
         )
 
-        xv, yv = chart.mapper.series_xy(series)
+        xv, yv = chart.series_xy(series)
         ax.plot(xv, yv, label=label, **kwargs)
 
         with np.errstate(invalid="ignore"):